backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.1.0
numpy==1.26.3
opencv-python-headless
matplotlib==3.8.2